+++ source/renderer/zettlr-renderer.js
@@ -24,7 +24,11 @@ const popup = require('./zettlr-popup.js')
 const ZettlrStatsView = require('./zettlr-stats-view.js')
 const ZettlrAttachments = require('./zettlr-attachments.js')
 
-const remote = require('electron').remote
+const { remote } = require('electron')
+const { clipboard } = require('electron')
+
+const { generateId } = require('../common/zettlr-helpers.js')
+
 const path = require('path')
 
 // Pull the poll-time from the data
@@ -173,6 +177,35 @@ class ZettlrRenderer {
 
     // And last but not least the zkn options
     this.getEditor().getEditor().setOption('zkn', global.config.get('zkn'))
+  }
+
+  /**
+   * Generates an ID based upon the configured pattern, writes it into the
+   * clipboard and then triggers the paste command on these webcontents.
+   * @return {void} Does not return.
+   */
+  genId () {
+    // First we need to backup the existing clipboard contents so that they are
+    // not lost during the operation.
+    let text = clipboard.readText()
+    let html = clipboard.readHTML()
+    let image = clipboard.readImage()
+    let rtf = clipboard.readRTF()
+
+    // Write an ID to the clipboard
+    clipboard.writeText(generateId(global.config.get('zkn.idGen')))
+    // Paste the ID
+    remote.getCurrentWebContents().paste()
+
+    // Now restore the clipboard's original contents
+    setTimeout((e) => {
+      clipboard.write({
+        'text': text,
+        'html': html,
+        'image': image,
+        'rtf': rtf
+      })
+    }, 10) // Why do a timeout? Because the paste event is asynchronous.
   }
 
   /**